requirements.txt
APScheduler==3.11.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzlocal==5.3.1
Werkzeug==3.1.3
